+++ app/epa/settings.py
@@ -19,7 +19,7 @@ from pathlib import Path
 import sys
 
 # Version number -> gets updated by bump-my-version, do not touch manually
-APP_VERSION_NUMBER = "2.1.1"
+APP_VERSION_NUMBER = "2.1.2"
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent