+++ app/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Fixed
 - Fix thermal efficiency value display for heat storage on results page [[#451](https://github.com/open-plan-tool/gui/pull/451)]
+- Small usability improvements like additional support messages or fixed translations [[#457](https://github.com/open-plan-tool/gui/pull/457)]
+- Remove faulty KPIs from results view [[#457](https://github.com/open-plan-tool/gui/pull/457)]
 
 ## [v2.1.1] – 2026-04-15
 ### Fixed